refactor: move logica de transacões para o lambda
