+++ src/functions/createTransaction.js
@@ -1,18 +1,143 @@
-const pagarme = require('pagarme');
+import pagarme from 'pagarme';
+
+import { parseAndFormatDateService, getFutureDate } from '../services/dateService';
+import { removeMaskService } from '../services/maskService';
 
 require('dotenv').config({
   path: `.env.${process.env.NODE_ENV}`,
 });
 
 export async function handler(event, context) {
-  return pagarme.client
-    .connect({ api_key: process.env.PAGARME_API_KEY })
-    .then(client => client.transactions.create(JSON.parse(event.body)))
-    .then(transaction => {
+  return createTransactions(JSON.parse(event.body))
+    .then(transactions => {
       return {
         statusCode: 200,
-        body: JSON.stringify(transaction),
+        body: JSON.stringify(transactions),
       };
     })
     .catch(error => ({ statusCode: 400, body: JSON.stringify(error) }));
+}
+
+const getProductsBySeller = bag => {
+  return bag.reduce((list, product) => {
+    const seller = product.seller.id;
+    if (list[seller]) {
+      list[seller] = [...list[seller], product];
+    } else {
+      list[seller] = [product];
+    }
+
+    return list;
+  }, {});
+};
+
+const customerPayload = personalData => {
+  return {
+    external_id: removeMaskService(personalData.cpf),
+    name: personalData.fullName,
+    type: 'individual',
+    country: 'br',
+    email: personalData.email,
+    documents: [
+      {
+        type: 'cpf',
+        number: removeMaskService(personalData.cpf),
+      },
+    ],
+    phone_numbers: [`+55${removeMaskService(personalData.phone)}`],
+    birthday: parseAndFormatDateService(personalData.birthdate, 'DD/MM/YYYY', 'YYYY-MM-DD'),
+  };
+};
+
+const billingPayload = (personalData, addressData) => {
+  return {
+    name: personalData.fullName,
+    address: {
+      country: 'br',
+      state: addressData.state,
+      city: addressData.city,
+      neighborhood: addressData.neighborhood,
+      street: addressData.street,
+      street_number: addressData.number,
+      zipcode: removeMaskService(addressData.zipCode),
+    },
+  };
+};
+
+const shippingPayload = (personalData, addressData) => {
+  return {
+    name: personalData.fullName,
+    fee: 0,
+    delivery_date: getFutureDate(3, 'YYYY-MM-DD'),
+    expedited: true,
+    address: {
+      country: 'br',
+      state: addressData.state,
+      city: addressData.city,
+      neighborhood: addressData.neighborhood,
+      street: addressData.street,
+      street_number: addressData.number,
+      zipcode: removeMaskService(addressData.zipCode),
+    },
+  };
+};
+
+async function createTransactions({ personalData, addressData, paymentData, bag }) {
+  const customer = customerPayload(personalData);
+  const billing = billingPayload(personalData, addressData);
+  const shipping = shippingPayload(personalData, addressData);
+
+  const productsBySeller = getProductsBySeller(bag);
+
+  const requests = Object.keys(productsBySeller).map(async seller => {
+    const products = productsBySeller[seller];
+
+    const items = products.map(item => ({
+      id: item.id,
+      title: item.name,
+      unit_price: item.price * 100,
+      quantity: item.quantity,
+      tangible: true,
+    }));
+
+    const amount = products.reduce((total, product) => total + product.price * 100, 0);
+
+    const body = {
+      amount,
+      card_number: paymentData.cardNumber,
+      card_cvv: paymentData.cvv,
+      card_expiration_date: removeMaskService(paymentData.expirationDate),
+      card_holder_name: paymentData.fullName,
+      customer,
+      billing,
+      shipping,
+      items,
+      split_rules: [
+        {
+          recipient_id: seller,
+          percentage: process.env.SELLER_PERCENTAGE,
+          liable: true,
+          charge_processing_fee: true,
+          charge_remainder: true,
+        },
+        {
+          recipient_id: process.env.PLATFORM_ID,
+          percentage: process.env.PLATFORM_PERCENTAGE,
+          liable: true,
+          charge_processing_fee: false,
+          charge_remainder: false,
+        },
+      ],
+    };
+
+    return pagarme.client
+      .connect({ api_key: process.env.PAGARME_API_KEY })
+      .then(client => client.transactions.create(body))
+      .then(transaction => {
+        return transaction;
+      })
+      .catch(error => error);
+  });
+
+  return Promise.all(requests);
 }
